+++ src/ralph/api/__init__.py
@@ -12,6 +12,7 @@ from ralph.conf import settings
 
 from .. import __version__
 from .auth import get_authenticated_user
+from .auth.oidc import AuthenticatedOidcUser
 from .auth.user import AuthenticatedUser
 from .routers import health, statements
 
@@ -50,7 +51,11 @@ if len(settings.RUNSERVER_CORS_ALLOW_ORIGINS) != 0:
         allow_credentials=True,
         allow_methods=["GET", "POST", "PUT", "DELETE"],
         allow_headers=[
-            "Authorization","User-Agent","Keep-Alive","Content-Type","X-Experience-API-Version"
+            "Authorization",
+            "User-Agent",
+            "Keep-Alive",
+            "Content-Type",
+            "X-Experience-API-Version",
         ],
     )
 app.include_router(statements.router)
@@ -62,7 +67,12 @@ async def whoami(
     user: AuthenticatedUser = Depends(get_authenticated_user),
 ) -> Dict[str, Any]:
     """Return the current user's username along with their scopes."""
-    return {
+    data = {
         "agent": user.agent.model_dump(mode="json", exclude_none=True),
         "scopes": user.scopes,
     }
+    if settings.LRS_EXTEND_AUTHORITY_TO_CLIENT_OWNERSHIP and isinstance(
+        user, AuthenticatedOidcUser
+    ):
+        data["client_agents"] = user.client_agents
+    return data